backend/pyproject.toml
[tool.poetry]
name = "concert-backend"
version = "0.1.0"
description = "Backend for Concert"
authors = ["arnu515"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sanic = {extras = ["ext"], version = "^22.9.1"}
prisma = "^0.7.1"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
